// ambulance-spa/src/app/model/pnForm.ts
export class PnForm {
    id: string;
    streetName: string;
    streetNumber: number;
    beginningDate: Date;
    endDate: Date;
    diagnoseCategory: number;
    diagnoseNumber: number;
    endDiagnose: number;
    patientBirthNumber: string;
    doctorId: number;
    tempAddressPSC: string;
    status: number;
}